Do not generate spilled coroutine fields multiple times
Otherwise, there will be VerifyError at runtime.
Previously, when suspend lambdas did not have invokeSuspend and
invokeSuspend$$forInline, there could be only two kinds of suspend
lambdas:
1) with state-machine and spilled variables
2) without both of them

Currently, however, we do not have the second kind. So, on
transformation of the lambda, we generate another field with the same
name.

This fix avoids it by memorizing FieldVisitors of transformed lambda,
so, when CoroutineTransformerMethodVisitor generates 'new' fields, it
reuses the old ones.